+++ src/rex.formbuilder/NEWS.rst
@@ -11,6 +11,8 @@ REX.FORMBUILDER Change Log
 * Fixed an issue where FireFox would highlight the float range constraint
   input boxes when you entered a float value, even though there was nothing
   wrong.
+* Fixed an issue that prevented the removal of pattern constraints on text
+  questions.
 
 
 5.6.2 (2016-02-29)